New our supporters page with provisional text
