+++ src/components/navbar/nav_links.js
@@ -17,7 +17,7 @@ export default [
       { intlId: "nav.ourData", sectionId: "", path: "/dashboard" },
       { intlId: "nav.ourVision", sectionId: "our-vision" },
       { intlId: "nav.ourTeam", sectionId: "our-team" },
-      { intlId: "nav.ourSupporters", sectionId: "our-supporters" },
+      { intlId: "nav.ourSupporters", sectionId: "", path: "/oursupporters" },
     ],
   },
   {
