+++ src/intl/en.json
@@ -492,6 +492,7 @@
   "team": {
     "supportersSection": {
       "header": "Our supporters",
+      "description": "The experience and determination of the organizations that contribute to Virufy represent the union of some of the best minds in the world and the most influential leaders in business, academia and society.",
       "supporterText": "Virufy is a <a target=\"_blank\" rel=\"noreferrer\" href=\"\" class=\"underline\"/>Stanford COVID-19 Response Innovation Lab</a> project and is also featured in <a target=\"_blank\" rel=\"noreferrer\" href=\"\" class=\"underline\"/>One Young World</a>"
     },
     "teamSection": {
