+++ src/intl/es.json
@@ -492,6 +492,7 @@
   "team": {
     "supportersSection": {
       "header": "Nuestros Patrocinadores",
+      "description": "La experiencia y determinación de las organizaciones que contribuyen a Virufy representan la unión de algunas de las mejores mentes del mundo y los líderes más influyentes del mundo empresarial, académico y social.",
       "supporterText": "Virufy es un proyecto del <a target=\"_blank\" rel=\"noreferrer\" href='https://responselab.stanford.edu/projects.html' class=\"underline\"/>Laboratorio de Innovación de Respuesta COVID-19 de Stanford</a> y también aparece en <a target=\"_blank\" rel=\"noreferrer\" href='https://www.oneyoungworld.com/' class=\"underline\"/>One Young World</a>"
     },
     "teamSection": {
