+++ src/intl/pt.json
@@ -492,6 +492,7 @@
   "team": {
     "supportersSection": {
       "header": "Os nossos apoiantes",
+      "description": "A experiência e determinação das organizações que contribuem para a Virufy representam a união de algumas das melhores cabeças do mundo e os líderes mais influentes nos negócios, na academia e na sociedade.",
       "supporterText": "Virufy is a <a target=\"_blank\" rel=\"noreferrer\" href=\"\" class=\"underline\"/>Stanford COVID-19 Response Innovation Lab</a> project and is also featured in <a target=\"_blank\" rel=\"noreferrer\" href=\"\" class=\"underline\"/>One Young World</a>"
     },
     "teamSection": {
