+++ src/pages/oursupporters.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import SEO from "../components/SEO";
+import Layout from "../components/Layout";
+import Supporters from "../components/oursupporters/Supporters";
+
+export default function OurSupporters() {
+  return (
+    <Layout>
+      <SEO title="Our Supporters | Virufy" />
+      <Supporters />
+    </Layout>
+  );
+}
